Adding InteropabilityTest_Engine
Migrating all push-pull-Compare methods to this new project.
