+++ InteropabilityTest_Engine/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="CompareNETObjects" version="4.65.0" targetFramework="net461" />
+</packages>
