+++ Test_Engine/Test_Engine.csproj
@@ -37,10 +37,6 @@
     <Reference Include="Diffing_oM">
       <HintPath>..\..\BHoM\Build\Diffing_oM.dll</HintPath>
     </Reference>
-    <Reference Include="KellermanSoftware.Compare-NET-Objects, Version=4.64.0.0, Culture=neutral, PublicKeyToken=d970ace04cc85217, processorArchitecture=MSIL">
-      <HintPath>..\packages\CompareNETObjects.4.64.0\lib\net46\KellermanSoftware.Compare-NET-Objects.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
     <Reference Include="netstandard" />
     <Reference Include="BHoM">
       <HintPath>..\..\BHoM\Build\BHoM.dll</HintPath>
@@ -126,6 +122,7 @@
     <Compile Include="Compute\CheckProjectStructure.cs" />
     <Compile Include="Compute\CheckReferences.cs" />
     <Compile Include="Compute\Check.cs" />
+    <Compile Include="Compute\StructuralResultsCheck.cs" />
     <Compile Include="Convert\ToBHoM.cs" />
     <Compile Include="Convert\ToSyntaxTree.cs" />
     <Compile Include="Convert\ToSpan.cs" />
@@ -141,7 +138,6 @@
     <Compile Include="Create\Span.cs" />
     <Compile Include="Modify\Merge.cs" />
     <Compile Include="Compute\PushPullCompare.cs" />
-    <Compile Include="Objects\EqualityComparers\AdapterIdComparer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\Checks\HasValidCopyright.cs" />
     <Compile Include="Query\Checks\CreateMethodFileNameMatchesReturnType.cs" />
@@ -171,7 +167,6 @@
     <Compile Include="Query\IsAdapterConstructor.cs" />
     <Compile Include="Query\IsDeprecated.cs" />
     <Compile Include="Query\IsEngineMethod.cs" />
-    <Compile Include="Query\IsEqual.cs" />
     <Compile Include="Query\IsImmutable.cs" />
     <Compile Include="Query\ObjectsAreIObject.cs" />
     <Compile Include="Query\Passes.cs" />
@@ -200,7 +195,6 @@
     <Compile Include="Query\GetNamespaceRoot.cs" />
     <Compile Include="Query\GetFileRoot.cs" />
     <Compile Include="Query\IsValidBHoMObject.cs" />
-    <Compile Include="Query\TestDataOfType.cs" />
     <Compile Include="Query\GetText.cs" />
   </ItemGroup>
   <ItemGroup>
