+++ Test_Engine/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="CompareNETObjects" version="4.64.0" targetFramework="net461" />
   <package id="Microsoft.CodeAnalysis.Analyzers" version="2.9.4" targetFramework="net461" developmentDependency="true" />
   <package id="Microsoft.CodeAnalysis.Common" version="3.4.0-beta2-final" targetFramework="net461" />
   <package id="Microsoft.CodeAnalysis.CSharp" version="3.4.0-beta2-final" targetFramework="net461" />
